Remove unused components and fixed the routing and created doctor page
