+++ client/src/Pages/Nurse.jsx
@@ -1,0 +1,9 @@
+
+
+export default function Nurse() {
+  return (
+    <div>
+      <h1 className="text-3xl text-slate-500">Add one more field in model.nurse so that total fields would be 18 i added the 16 field right now... </h1>
+    </div>
+  )
+}
