+++ client/src/Pages/Patients.jsx
@@ -2,7 +2,6 @@ import { useState } from 'react'
 import Patient_detail from './Patient_detail.jsx'
 import CarePlan from './Careplandetail.jsx'
 import Appointment from './Appointdetail.jsx'
-import Layout from './Layout.jsx'
 
 export default function Patients(){
 
@@ -25,7 +24,6 @@ export default function Patients(){
   }
 
   return (
-    <Layout>
     <div className='flex flex-col'>
       <div className='flex mb-2'>
         <button
@@ -55,7 +53,6 @@ export default function Patients(){
       </div>
       {activeContent}
     </div>
-    </Layout>
   )
 }
